bugfixes direction

+++ anglehelper/const.go
@@ -3,6 +3,6 @@ package anglehelper
 import "math"
 
 const (
-	TwoPi        = math.Pi * 2.0
+	TwoPi        = float64(math.Pi) * 2.0
 	ReverseTwoPi = 1.0 / TwoPi
 )
